+++ substrate/runtime/staking/src/lib.rs
@@ -61,91 +61,18 @@ use primitives::traits::{Zero, One, Bounded, RefInto, SimpleArithmetic, Executab
 	As, Lookup};
 use primitives::generic::{Member, MaybeSerializeDebug};
 
+pub mod address;
 mod contract;
 #[cfg(test)]
 mod mock;
+#[cfg(test)]
+mod tests;
 
 /// Number of account IDs stored per enum set.
 const ENUM_SET_SIZE: usize = 64;
 
 /// The byte to identify intention to reclaim an existing account index.
 const RECLAIM_INDEX_MAGIC: usize = 0x69;
-
-pub mod address {
-	use super::{Member, Slicable, As, Input};
-
-	/// A vetted and verified extrinsic from the external world.
-	#[derive(PartialEq, Eq, Clone)]
-	#[cfg_attr(feature = "std", derive(Serialize, Debug))]
-	pub enum Address<AccountId, AccountIndex> where
-	 	AccountId: Member,
-	 	AccountIndex: Member,
-	{
-		/// It's an account ID (pubkey).
-		Id(AccountId),
-		/// It's an account index.
-		Index(AccountIndex),
-	}
-
-	impl<AccountId, AccountIndex> From<AccountId> for Address<AccountId, AccountIndex> where
-	 	AccountId: Member,
-	 	AccountIndex: Member,
-	{
-		fn from(a: AccountId) -> Self {
-			Address::Id(a)
-		}
-	}
-
-	impl<AccountId, AccountIndex> Slicable for Address<AccountId, AccountIndex> where
-		AccountId: Member + Slicable,
-		AccountIndex: Member + Slicable + PartialOrd<AccountIndex> + Ord + As<u32> + As<u16> + As<u8> + Copy,
-	{
-		fn decode<I: Input>(input: &mut I) -> Option<Self> {
-			match input.read_byte()? {
-				255 => Some(Address::Id(Slicable::decode(input)?)),
-				254 => Some(Address::Index(Slicable::decode(input)?)),
-				253 => Some(Address::Index(As::sa(u32::decode(input)?))),
-				252 => Some(Address::Index(As::sa(u16::decode(input)?))),
-				x => Some(Address::Index(As::sa(x))),
-			}
-		}
-
-		fn encode(&self) -> Vec<u8> {
-			let mut v = Vec::new();
-
-			match *self {
-				Address::Id(ref i) => {
-					v.push(255);
-					i.using_encoded(|s| v.extend(s));
-				}
-				Address::Index(i) if i > As::sa(0xffffffffu32) => {
-					v.push(254);
-					i.using_encoded(|s| v.extend(s));
-				}
-				Address::Index(i) if i > As::sa(0xffffu32) => {
-					v.push(253);
-					As::<u32>::as_(i).using_encoded(|s| v.extend(s));
-				}
-				Address::Index(i) if i >= As::sa(252u32) => {
-					v.push(252);
-					As::<u16>::as_(i).using_encoded(|s| v.extend(s));
-				}
-				Address::Index(i) => v.push(As::<u8>::as_(i)),
-			}
-
-			v
-		}
-	}
-
-	impl<AccountId, AccountIndex> Default for Address<AccountId, AccountIndex> where
-		AccountId: Member + Default,
-		AccountIndex: Member,
-	{
-		fn default() -> Self {
-			Address::Id(Default::default())
-		}
-	}
-}
 
 pub type Address<T> = address::Address<<T as system::Trait>::AccountId, <T as Trait>::AccountIndex>;
 
@@ -963,339 +890,5 @@ impl<T: Trait> primitives::BuildExternalities for GenesisConfig<T> {
 			r.insert(twox_128(&<FreeBalance<T>>::key_for(who)).to_vec(), value.encode());
 		}
 		r
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use runtime_io::with_externalities;
-	use mock::*;
-
-	#[test]
-	fn indexing_lookup_should_work() {
-		with_externalities(&mut new_test_ext(10, 1, 2, 0, true), || {
-			assert_eq!(Staking::lookup_index(0), Some(1));
-			assert_eq!(Staking::lookup_index(1), Some(2));
-			assert_eq!(Staking::lookup_index(2), Some(3));
-			assert_eq!(Staking::lookup_index(3), Some(4));
-			assert_eq!(Staking::lookup_index(4), None);
-		});
-	}
-
-	#[test]
-	fn default_indexing_on_new_accounts_should_work() {
-		with_externalities(&mut new_test_ext(10, 1, 2, 0, true), || {
-			assert_eq!(Staking::lookup_index(4), None);
-			assert_ok!(Staking::transfer(&1, 5.into(), 10));
-			assert_eq!(Staking::lookup_index(4), Some(5));
-		});
-	}
-
-	#[test]
-	fn dust_account_removal_should_work() {
-		with_externalities(&mut new_test_ext(256 * 10, 1, 2, 0, true), || {
-			assert_eq!(Staking::balance(&2), 256 * 20);
-			assert_ok!(Staking::transfer(&2, 5.into(), 256 * 10 + 1));	// index 1 (account 2) becomes zombie
-			assert_eq!(Staking::balance(&2), 0);
-			assert_eq!(Staking::balance(&5), 256 * 10 + 1);
-		});
-	}
-
-	#[test]
-	fn reclaim_indexing_on_new_accounts_should_work() {
-		with_externalities(&mut new_test_ext(256 * 1, 1, 2, 0, true), || {
-			assert_eq!(Staking::lookup_index(1), Some(2));
-			assert_eq!(Staking::lookup_index(4), None);
-			assert_eq!(Staking::balance(&2), 256 * 20);
-			assert_ok!(Staking::transfer(&2, 5.into(), 256 * 20));	// account 2 becomes zombie freeing index 1 for reclaim)
-			assert_eq!(Staking::balance(&2), 0);
-			assert_ok!(Staking::transfer(&5, 6.into(), 256 * 1 + 0x69));	// account 6 takes index 1.
-			assert_eq!(Staking::balance(&6), 256 * 1 + 0x69);
-			assert_eq!(Staking::lookup_index(1), Some(6));
-		});
-	}
-
-	#[test]
-	fn staking_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 2, 0, true), || {
-			assert_eq!(Staking::era_length(), 2);
-			assert_eq!(Staking::validator_count(), 2);
-			assert_eq!(Staking::bonding_duration(), 3);
-			assert_eq!(Session::validators(), vec![10, 20]);
-
-			// Block 1: Add three validators. No obvious change.
-			System::set_block_number(1);
-			assert_ok!(Staking::stake(&1));
-			assert_ok!(Staking::stake(&2));
-			assert_ok!(Staking::stake(&4));
-			Staking::check_new_era();
-			assert_eq!(Session::validators(), vec![10, 20]);
-
-			// Block 2: New validator set now.
-			System::set_block_number(2);
-			Staking::check_new_era();
-			assert_eq!(Session::validators(), vec![4, 2]);
-
-			// Block 3: Unstake highest, introduce another staker. No change yet.
-			System::set_block_number(3);
-			assert_ok!(Staking::stake(&3));
-			assert_ok!(Staking::unstake(&4));
-			Staking::check_new_era();
-
-			// Block 4: New era - validators change.
-			System::set_block_number(4);
-			Staking::check_new_era();
-			assert_eq!(Session::validators(), vec![3, 2]);
-
-			// Block 5: Transfer stake from highest to lowest. No change yet.
-			System::set_block_number(5);
-			assert_ok!(Staking::transfer(&4, 1.into(), 40));
-			Staking::check_new_era();
-
-			// Block 6: Lowest now validator.
-			System::set_block_number(6);
-			Staking::check_new_era();
-			assert_eq!(Session::validators(), vec![1, 3]);
-
-			// Block 7: Unstake three. No change yet.
-			System::set_block_number(7);
-			assert_ok!(Staking::unstake(&3));
-			Staking::check_new_era();
-			assert_eq!(Session::validators(), vec![1, 3]);
-
-			// Block 8: Back to one and two.
-			System::set_block_number(8);
-			Staking::check_new_era();
-			assert_eq!(Session::validators(), vec![1, 2]);
-		});
-	}
-
-	#[test]
-	fn staking_eras_work() {
-		with_externalities(&mut new_test_ext(0, 1, 2, 0, true), || {
-			assert_eq!(Staking::era_length(), 2);
-			assert_eq!(Staking::sessions_per_era(), 2);
-			assert_eq!(Staking::last_era_length_change(), 0);
-			assert_eq!(Staking::current_era(), 0);
-
-			// Block 1: No change.
-			System::set_block_number(1);
-			Staking::check_new_era();
-			assert_eq!(Staking::sessions_per_era(), 2);
-			assert_eq!(Staking::last_era_length_change(), 0);
-			assert_eq!(Staking::current_era(), 0);
-
-			// Block 2: Simple era change.
-			System::set_block_number(2);
-			Staking::check_new_era();
-			assert_eq!(Staking::sessions_per_era(), 2);
-			assert_eq!(Staking::last_era_length_change(), 0);
-			assert_eq!(Staking::current_era(), 1);
-
-			// Block 3: Schedule an era length change; no visible changes.
-			System::set_block_number(3);
-			assert_ok!(Staking::set_sessions_per_era(3));
-			Staking::check_new_era();
-			assert_eq!(Staking::sessions_per_era(), 2);
-			assert_eq!(Staking::last_era_length_change(), 0);
-			assert_eq!(Staking::current_era(), 1);
-
-			// Block 4: Era change kicks in.
-			System::set_block_number(4);
-			Staking::check_new_era();
-			assert_eq!(Staking::sessions_per_era(), 3);
-			assert_eq!(Staking::last_era_length_change(), 4);
-			assert_eq!(Staking::current_era(), 2);
-
-			// Block 5: No change.
-			System::set_block_number(5);
-			Staking::check_new_era();
-			assert_eq!(Staking::sessions_per_era(), 3);
-			assert_eq!(Staking::last_era_length_change(), 4);
-			assert_eq!(Staking::current_era(), 2);
-
-			// Block 6: No change.
-			System::set_block_number(6);
-			Staking::check_new_era();
-			assert_eq!(Staking::sessions_per_era(), 3);
-			assert_eq!(Staking::last_era_length_change(), 4);
-			assert_eq!(Staking::current_era(), 2);
-
-			// Block 7: Era increment.
-			System::set_block_number(7);
-			Staking::check_new_era();
-			assert_eq!(Staking::sessions_per_era(), 3);
-			assert_eq!(Staking::last_era_length_change(), 4);
-			assert_eq!(Staking::current_era(), 3);
-		});
-	}
-
-	#[test]
-	fn staking_balance_works() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 42);
-			assert_eq!(Staking::free_balance(&1), 42);
-			assert_eq!(Staking::reserved_balance(&1), 0);
-			assert_eq!(Staking::balance(&1), 42);
-			assert_eq!(Staking::free_balance(&2), 0);
-			assert_eq!(Staking::reserved_balance(&2), 0);
-			assert_eq!(Staking::balance(&2), 0);
-		});
-	}
-
-	#[test]
-	fn staking_balance_transfer_works() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::transfer(&1, 2.into(), 69));
-			assert_eq!(Staking::balance(&1), 42);
-			assert_eq!(Staking::balance(&2), 69);
-		});
-	}
-
-	#[test]
-	fn staking_balance_transfer_when_bonded_should_not_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::stake(&1));
-			assert_noop!(Staking::transfer(&1, 2.into(), 69), "bondage too high to send value");
-		});
-	}
-
-	#[test]
-	fn reserving_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-
-			assert_eq!(Staking::balance(&1), 111);
-			assert_eq!(Staking::free_balance(&1), 111);
-			assert_eq!(Staking::reserved_balance(&1), 0);
-
-			assert_ok!(Staking::reserve_balance(&1, 69));
-
-			assert_eq!(Staking::balance(&1), 111);
-			assert_eq!(Staking::free_balance(&1), 42);
-			assert_eq!(Staking::reserved_balance(&1), 69);
-		});
-	}
-
-	#[test]
-	fn staking_balance_transfer_when_reserved_should_not_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::reserve_balance(&1, 69));
-			assert_noop!(Staking::transfer(&1, 2.into(), 69), "balance too low to send value");
-		});
-	}
-
-	#[test]
-	fn deducting_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::deduct_unbonded(&1, 69));
-			assert_eq!(Staking::free_balance(&1), 42);
-		});
-	}
-
-	#[test]
-	fn deducting_balance_when_bonded_should_not_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			<Bondage<Test>>::insert(1, 2);
-			System::set_block_number(1);
-			assert_eq!(Staking::unlock_block(&1), LockStatus::LockedUntil(2));
-			assert_noop!(Staking::deduct_unbonded(&1, 69), "not enough liquid funds");
-		});
-	}
-
-	#[test]
-	fn refunding_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 42);
-			Staking::refund(&1, 69);
-			assert_eq!(Staking::free_balance(&1), 111);
-		});
-	}
-
-	#[test]
-	fn slashing_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::reserve_balance(&1, 69));
-			assert_ok!(Staking::slash(&1, 69));
-			assert_eq!(Staking::free_balance(&1), 0);
-			assert_eq!(Staking::reserved_balance(&1), 42);
-		});
-	}
-
-	#[test]
-	fn slashing_incomplete_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 42);
-			assert_ok!(Staking::reserve_balance(&1, 21));
-			assert_err!(Staking::slash(&1, 69), "not enough funds");
-			assert_eq!(Staking::free_balance(&1), 0);
-			assert_eq!(Staking::reserved_balance(&1), 0);
-		});
-	}
-
-	#[test]
-	fn unreserving_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::reserve_balance(&1, 111));
-			Staking::unreserve_balance(&1, 42);
-			assert_eq!(Staking::reserved_balance(&1), 69);
-			assert_eq!(Staking::free_balance(&1), 42);
-		});
-	}
-
-	#[test]
-	fn slashing_reserved_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::reserve_balance(&1, 111));
-			assert_ok!(Staking::slash_reserved(&1, 42));
-			assert_eq!(Staking::reserved_balance(&1), 69);
-			assert_eq!(Staking::free_balance(&1), 0);
-		});
-	}
-
-	#[test]
-	fn slashing_incomplete_reserved_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::reserve_balance(&1, 42));
-			assert_err!(Staking::slash_reserved(&1, 69), "not enough (reserved) funds");
-			assert_eq!(Staking::free_balance(&1), 69);
-			assert_eq!(Staking::reserved_balance(&1), 0);
-		});
-	}
-
-	#[test]
-	fn transferring_reserved_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::reserve_balance(&1, 111));
-			assert_ok!(Staking::transfer_reserved_balance(&1, &2, 42));
-			assert_eq!(Staking::reserved_balance(&1), 69);
-			assert_eq!(Staking::free_balance(&1), 0);
-			assert_eq!(Staking::reserved_balance(&2), 0);
-			assert_eq!(Staking::free_balance(&2), 42);
-		});
-	}
-
-	#[test]
-	fn transferring_incomplete_reserved_balance_should_work() {
-		with_externalities(&mut new_test_ext(0, 1, 3, 1, false), || {
-			<FreeBalance<Test>>::insert(1, 111);
-			assert_ok!(Staking::reserve_balance(&1, 42));
-			assert_err!(Staking::transfer_reserved_balance(&1, &2, 69), "not enough (reserved) funds");
-			assert_eq!(Staking::reserved_balance(&1), 0);
-			assert_eq!(Staking::free_balance(&1), 69);
-			assert_eq!(Staking::reserved_balance(&2), 0);
-			assert_eq!(Staking::free_balance(&2), 42);
-		});
 	}
 }
